criando estrutura inicial de DTO, repository, service e controller de Category

+++ product-api/src/main/java/com/alanjsantos/productapi/model/Category.java
@@ -2,6 +2,7 @@ package com.alanjsantos.productapi.model;
 
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -15,7 +16,7 @@ public class Category {
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    private Long id;
+    private Integer id;
 
     @Column(nullable = false)
     private String description;
